+++ lib/src/actor/actor_message_receiver_mixin.dart
@@ -1,6 +1,7 @@
 part of theater.actor;
 
 mixin ActorMessageReceiverMixin<P extends ActorProperties> on ActorContext<P> {
+  /// Sets [handler] to handle all messages of type [T] that received actor.
   StreamSubscription<MailboxMessage> receive<T>(
       Future<MessageResult?> Function(T) handler) {
     var subscription = _messageController.stream.listen((message) async {