+++ lib/src/actor/router_actor/pool_router_actor/pool_router_actor_context.dart
@@ -4,7 +4,8 @@ part of theater.actor;
 ///
 /// Receive messages from other actors, create a pool of child actors and control their life cycle.
 class PoolRouterActorContext
-    extends RouterActorContext<PoolRouterActorProperties> {
+    extends RouterActorContext<PoolRouterActorProperties>
+    with UserActorContextMixin<PoolRouterActorProperties> {
   final Map<ActorCell, int> _loadingMap = {};
 
   PoolRouterActorContext(