+++ lib/src/dispatch/ref/ref_register/ref_register.dart
@@ -1,0 +1,7 @@
+part of theater.dispatch;
+
+abstract class RefRegister<T extends Ref> {
+  final List<T> _refs = [];
+
+  List<T> get refs => List.unmodifiable(_refs);
+}