+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Duplicates each method of sending a message to another actor. There is a version with a subscription to the state of the message (and to reply to messages) and without a subscription. It was done because it is not always necessary to monitor the status or response of a message, but this process creates additional traffic and degrades the speed of sending the application, the performance of the application. Therefore, use the subscription only when you need it.
 
+- Adds new way to get ref to actor. Now with ActorContext and ActorSystem, you can get the link to other actor.
+
 - Improves scheduler. In addition to repeatedly actions, there are one shot actions.
   - CancellationToken replace to RepeatedlyActionToken for repeatedly actions and OneShotActionToken for one shot actions.
   - For OneShotActionToken and RepeatedlyActionToken add the ability to receive links to them and transfer them to other actors (isolates).